+++ docs/source/conf.py
@@ -69,6 +69,7 @@ exclude_patterns = []
 #
 html_theme = 'pydata_sphinx_theme'
 html_last_updated_fmt = ""
+html_show_sourcelink = False
 html_theme_options = {
     "logo": {
         "image_light": "_static/RockVerse_logo_model3_for_white_background_facecolor_transparent_True.png",